minor changes and included deleting like document when tweets gets deleted

+++ mutation/tweet_mutation.py
@@ -85,6 +85,7 @@ class TweetMutation:
             return GeneralResponse(
                 msg="Authentication Failed: Invalid Token", success=False
             )
+        _ = like_collection.delete_one({"tweet_id": tweet_id})
         deleted_obj = tweet_collection.delete_one(
             {
                 "user_id": ObjectId(user_data["response"]["user_id"]),
@@ -102,9 +103,11 @@ class TweetMutation:
             return GeneralResponse(
                 msg="Authentication Failed: Invalid Token", success=False
             )
-        deleted_obj = tweet_collection.delete_many(
-            {"user_id": ObjectId(user_data["response"]["user_id"])}
-        )
+        user_id = ObjectId(user_data["response"]["user_id"])
+        tweet_objs = tweet_collection.find({"user_id": user_id})
+        tweetIds = [tweet["_id"] for tweet in tweet_objs]
+        _ = like_collection.delete_many({"tweet_id": {"$in": tweetIds}})
+        deleted_obj = tweet_collection.delete_many({"user_id": user_id})
         if not deleted_obj.deleted_count > 0:
             return GeneralResponse(msg="Delete Tweets Failed", success=False)
         return GeneralResponse(msg="Deleted All User Tweets", success=True)
